refactor(ui): extract ActivityDataTab component into its own file
