+++ app/src/app/[lng]/data/[step]/subsector-drawer.tsx
@@ -35,6 +35,7 @@ import { RefObject, useEffect, useState } from "react";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { Trans } from "react-i18next/TransWithoutContext";
 import { MdError } from "react-icons/md";
+import { ActivityDataTab } from "./ActivityDataTab";
 
 type Inputs = {
   valueType: string;
@@ -74,83 +75,19 @@ const defaultValues: Inputs = {
   gridSourceReference: "",
 };
 
-const activityDataUnits = ["kWh", "Unit1", "Unit2", "Unit3"];
-const emissionFactorTypes = [
-  "Local",
-  "Regional",
-  "National",
-  "IPCC",
-  "Add custom",
-];
-
-function ActivityDataTab({
+function DirectMeasureForm({
   t,
   register,
   errors,
-  prefix,
+  className,
 }: {
   t: TFunction;
   register: Function;
   errors: Record<string, any>;
-  prefix: string;
+  className?: string;
 }) {
   return (
-    <TabPanel px={0.5}>
-      <HStack spacing={4} mb={12} className="items-start">
-        <FormControl isInvalid={!!errors[prefix + "ActivityDataAmount"]}>
-          <FormLabel>
-            {t("activity-data-amount")}{" "}
-            <Tooltip
-              hasArrow
-              label={t("value-types-tooltip")}
-              placement="bottom-start"
-            >
-              <InfoOutlineIcon mt={-0.5} color="contentTertiary" />
-            </Tooltip>
-          </FormLabel>
-          <InputGroup>
-            <NumberInput defaultValue={0} w="full">
-              <NumberInputField
-                placeholder={t("activity-data-amount-placeholder")}
-                borderRightRadius={0}
-                {...register(prefix + "ActivityDataAmount", {
-                  required: t("activity-data-amount-required"),
-                })}
-              />
-            </NumberInput>
-            <InputRightAddon
-              className="border-l-2"
-              pl={4}
-              pr={0}
-              bgColor="white"
-            >
-              <Select
-                variant="unstyled"
-                {...register(prefix + "ActivityDataUnit")}
-              >
-                {activityDataUnits.map((unit) => (
-                  <option key={unit} value={unit}>
-                    {unit}
-                  </option>
-                ))}
-              </Select>
-            </InputRightAddon>
-          </InputGroup>
-          <FormErrorMessage>
-            {errors[prefix + "ActivityDataAmount"]?.message}
-          </FormErrorMessage>
-        </FormControl>
-        <FormControl>
-          <FormLabel>{t("emission-factor-type")}</FormLabel>
-          <Select {...register(prefix + "EmissionFactorType")}>
-            {emissionFactorTypes.map((type) => (
-              <option key={type} value={type}>
-                {type}
-              </option>
-            ))}
-          </Select>
-        </FormControl>
-      </HStack>
+    <Box className={className}>
       <Heading size="sm" mb={4} className="font-normal">
         {t("emission-factors-values")}{" "}
         <Tooltip
@@ -164,13 +101,13 @@ function ActivityDataTab({
       <HStack spacing={4} mb={5}>
         <FormControl>
           <FormLabel color="contentTertiary">
-            {t("co2-emission-factor")}
+            {t("co2-emissions-value")}
           </FormLabel>
           <InputGroup>
             <NumberInput defaultValue={0} min={0}>
               <NumberInputField
                 borderRightRadius={0}
-                {...register(prefix + "Co2EmissionFactor")}
+                {...register("directCo2Emissions")}
                 bgColor="backgroundNeutral"
               />
             </NumberInput>
@@ -178,20 +115,20 @@ function ActivityDataTab({
               bgColor="backgroundNeutral"
               color="contentTertiary"
             >
-              CO2/kWh
+              tCO2e
             </InputRightAddon>
           </InputGroup>
           <FormHelperText>&nbsp;</FormHelperText>
         </FormControl>
         <FormControl>
           <FormLabel color="contentTertiary">
-            {t("n2o-emission-factor")}
+            {t("ch4-emissions-value")}
           </FormLabel>
           <InputGroup>
             <NumberInput defaultValue={0} min={0}>
               <NumberInputField
                 borderRightRadius={0}
-                {...register(prefix + "N2oEmissionFactor")}
+                {...register("directCh4Emissions")}
                 bgColor="backgroundNeutral"
               />
             </NumberInput>
@@ -199,7 +136,7 @@ function ActivityDataTab({
               bgColor="backgroundNeutral"
               color="contentTertiary"
             >
-              N2O/kWh
+              tCH4e
             </InputRightAddon>
           </InputGroup>
           <FormHelperText color="contentTertiary">
@@ -208,13 +145,13 @@ function ActivityDataTab({
         </FormControl>
         <FormControl>
           <FormLabel color="contentTertiary">
-            {t("ch4-emission-factor")}
+            {t("n2o-emissions-value")}
           </FormLabel>
           <InputGroup>
             <NumberInput defaultValue={0} min={0}>
               <NumberInputField
                 borderRightRadius={0}
-                {...register(prefix + "Ch4EmissionFactor")}
+                {...register("directN2oEmissions")}
                 bgColor="backgroundNeutral"
               />
             </NumberInput>
@@ -222,7 +159,7 @@ function ActivityDataTab({
               bgColor="backgroundNeutral"
               color="contentTertiary"
             >
-              CH4/kWh
+              tN2Oe
             </InputRightAddon>
           </InputGroup>
           <FormHelperText color="contentTertiary">
@@ -230,35 +167,7 @@ function ActivityDataTab({
           </FormHelperText>
         </FormControl>
       </HStack>
-      <HStack className="items-start" mb={5}>
-        <InfoOutlineIcon mt={1} color="contentLink" />
-        <Text color="contentTertiary">{t("emissions-factor-details")}</Text>
-      </HStack>
-      <FormControl isInvalid={!!errors[prefix + "SourceReference"]} mb={12}>
-        <FormLabel>{t("source-reference")}</FormLabel>
-        <Textarea
-          placeholder={t("source-reference-placeholder")}
-          {...register(prefix + "SourceReference", {
-            required: t("source-reference-required"),
-          })}
-        />
-        <FormErrorMessage>
-          {errors[prefix + "SourceReference"]?.message}
-        </FormErrorMessage>
-      </FormControl>
-      <HStack className="items-start" mb={13}>
-        <InfoOutlineIcon mt={1} color="contentLink" />
-        <Text color="contentTertiary">
-          <Trans
-            t={t}
-            i18nKey="calculations-details"
-            values={{ gwpValue: 1337 }}
-          >
-            All calculations consider a <b>GWP value of X</b>.
-          </Trans>
-        </Text>
-      </HStack>
-    </TabPanel>
+    </Box>
   );
 }
 
@@ -400,7 +309,7 @@ export function SubsectorDrawer({
                   </RadioButton>
                 </HStack>
                 <Box
-                  className={`${valueType ? undefined : "invisible"} space-y-6`}
+                  className={`${valueType ? undefined : "hidden"} space-y-6`}
                 >
                   <Heading size="sm">
                     {t("select-methodology")}{" "}
@@ -430,7 +339,7 @@ export function SubsectorDrawer({
                       {t("direct-measure")}
                     </RadioButton>
                   </HStack>
-                  <Tabs className={methodology ? undefined : "invisible"}>
+                  <Tabs className={methodology == "activity-data" ? undefined : "hidden"}>
                     <TabList>
                       <Tab>
                         {t("fuel-combustion")}{" "}
@@ -470,11 +379,12 @@ export function SubsectorDrawer({
                       />
                     </TabPanels>
                   </Tabs>
+                  <DirectMeasureForm className={methodology == "direct-measure" ? undefined : "hidden"} t={t} register={register} errors={errors} />
                 </Box>
               </form>
             </DrawerBody>
           )}
-          <Stack w="full" py={6} className="drop-shadow-top border-t-2">
+          <Stack w="full" py={6} className="drop-shadow-top border-t-2" mt={6}>
             <Button
               onClick={handleSubmit(onSubmit)}
               isDisabled={!isSubmitEnabled}
